data_collector/requirements.txt
psycopg2-binary
requests
filelock